WIP volume rendering
Relized that colors, and densities will have dimensions (B,R,D) where B
is the batch size, R is the granularity of the shot ray, and D is the
dimension of the color space and density space (i.e. 3 and 1). I think
that several of the functions (volume rendering) have not taken account
for this extra dimension (R).