Tester done, however improvements could be needed in the future
I went a little off when it comes to the code standard that I've tried
to implement. However one should not worry to much about that for now.
Normalization (of the RGB values) could be the move since matplotlib can
interpret [0.,1.] as proper RGB values.

Otherwise the next move should be to try to get this naive model to
converge during training.